requirements.txt
certifi==2018.10.15
cloudinary==1.14.0
dj-database-url==0.5.0
Django==2.1.3
djangorestframework==3.9.0
mock==2.0.0
pbr==5.1.0
pytz==2018.7
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.1
